Avoid non-preemptible retries when '--preemptible'  (#18)
* Avoid non-preemptible retries

when '--preemptible' is set. We also enforce '--max_non_preemptible_tries'
to be set to 0 when preemptible option is selected.

* simplify flags by introducing --attempts

By replacing --max_non_preemptible_tries and --max_preemptible_tries with --attempts we make our flags more straightforward.
Note, we need to update our docs after this change is released:
https://cloud.google.com/genomics/docs/tutorials/deepvariant

* Make adding --attemtps flag backward compatible

We keep the current two flags and their default values:
 * --max_preemptible_tries
 * --max_non_preemptible_tries
we also set the default value of the new --attempts flags 0.

The underlying logic will match fixed logic, ie when --preemptible
is set, all retires will be using preemptible machines. This way we
avoid any unexpected cost associated with using non-preemptible
machines.

As issue #22 indicates, we will deprecate these two flags in the next
release and set meaningful default value for --attempts flag. Meanwhile,
we are showing a warning when these two flags are being used and also we
do not allow to use them at the same time when --attempts is being used.